+++ packages/portal-proto/src/features/proteinpaint/IDCViewerWrapper.tsx
@@ -8,6 +8,8 @@ import {
   convertFilterToGqlFilter, // added: convert Operation -> GqlOperation
   useCurrentCohortCounts, // added hook to get cohort case count
 } from "@gff/core";
+// import the new split components
+import TableView from "./TableView";
 
 /**
  * Cleaned IDCViewerWrapper:
@@ -348,291 +350,13 @@ const IDCViewerWrapper: FC = () => {
 
       <div style={{ marginTop: 12 }}>
         <div style={{ maxHeight: 460, overflow: "auto" }}>
-          {mappings.length === 0 ? (
-            <div>Loading GDC - IDC mappings</div>
-          ) : (
-            <>
-              {/* replace existing small style block with sticky header styles */}
-              <style>
-                {`
-                  .idc-mapping-table {
-                    border-collapse: collapse;
-                  }
-                  /* keep existing hover behavior */
-                  .idc-mapping-table tbody tr:hover {
-                    background: #fff9e6 !important;
-                  }
-
-                  /* sticky header */
-                  .idc-mapping-table thead th {
-                    position: sticky;
-                    top: 0;
-                    background: #fafafa; /* match table background for visual continuity */
-                    z-index: 4;
-                    padding: 6px 8px;
-                    text-align: left;
-                    border-bottom: 1px solid #ddd;
-                  }
-
-                  /* ensure header content doesn't overlap when sticky */
-                  .idc-mapping-table thead tr {
-                    /* keep header row height consistent */
-                  }
-                `}
-              </style>
-
-              <table
-                className="idc-mapping-table"
-                style={{
-                  width: "100%",
-                  borderCollapse: "collapse",
-                  fontSize: 12,
-                  background: "#fafafa",
-                  borderRadius: 6,
-                  tableLayout: "fixed",
-                }}
-              >
-                <colgroup>
-                  <col style={{ width: "10%" }} /> {/* GDC caseId */}
-                  <col style={{ width: "30%" }} />{" "}
-                  {/* StudyInstanceUID / count + arrow */}
-                  <col style={{ width: "15%" }} /> {/* StudyDate */}
-                  <col style={{ width: "25%" }} /> {/* StudyDescription */}
-                  <col style={{ width: "10%" }} /> {/* WSI link */}
-                  <col style={{ width: "10%" }} /> {/* Radiology Link */}
-                </colgroup>
-                <thead>
-                  <tr
-                    style={{
-                      textAlign: "left",
-                      borderBottom: "1px solid #ddd",
-                    }}
-                  >
-                    <th style={{ padding: "6px 8px" }}>GDC caseId</th>
-                    <th style={{ padding: "6px 8px" }}>
-                      IDC StudyInstanceUUID
-                    </th>
-                    <th style={{ padding: "6px 8px" }}>StudyDate</th>
-                    <th style={{ padding: "6px 8px" }}>StudyDescription</th>
-                    <th style={{ padding: "6px 8px" }}>WSI link</th>
-                    <th style={{ padding: "6px 8px" }}>Radiology Link</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {mappings.map((m, idx) => {
-                    const caseId =
-                      m.gdcCase.submitter_id ??
-                      m.gdcCase.case_id ??
-                      m.gdcCase.case_uuid ??
-                      "(no id)";
-
-                    // --- REPLACED: build studiesMap (group by StudyInstanceUID) ---
-                    // Use precise flags:
-                    //  - hasRadiology: modalities CT, MRI, PET
-                    //  - hasWSI: modality SM (whole-slide images)
-                    const studiesMap = new Map<
-                      string,
-                      {
-                        StudyInstanceUID: string | null;
-                        series: {
-                          SeriesInstanceUID?: string | null;
-                          Modality?: string | null;
-                        }[];
-                        hasRadiology: boolean;
-                        hasWSI: boolean;
-                        StudyDate?: string | null;
-                        StudyDescription?: string | null;
-                      }
-                    >();
-
-                    if (Array.isArray(m.matches)) {
-                      for (const row of m.matches) {
-                        const studyId = row?.StudyInstanceUID ?? null;
-                        const key = studyId ?? "__NO_STUDY__";
-                        let existing = studiesMap.get(key);
-                        if (!existing) {
-                          existing = {
-                            StudyInstanceUID: studyId,
-                            series: [],
-                            hasRadiology: false,
-                            hasWSI: false,
-                            StudyDate: row?.StudyDate ?? null,
-                            StudyDescription: row?.StudyDescription ?? null,
-                          };
-                          studiesMap.set(key, existing);
-                        }
-                        // collect series info
-                        existing.series.push({
-                          SeriesInstanceUID: row?.SeriesInstanceUID ?? null,
-                          Modality: row?.Modality ?? null,
-                        });
-                        // prefer first-seen non-null StudyDate / StudyDescription
-                        if (!existing.StudyDate && row?.StudyDate)
-                          existing.StudyDate = row.StudyDate;
-                        if (!existing.StudyDescription && row?.StudyDescription)
-                          existing.StudyDescription = row.StudyDescription;
-                        // precise modality classification
-                        const mod = (row?.Modality ?? "")
-                          .toString()
-                          .trim()
-                          .toUpperCase();
-                        if (mod === "SM") {
-                          existing.hasWSI = true;
-                        } else if (
-                          mod === "CT" ||
-                          mod === "MRI" ||
-                          mod === "PET"
-                        ) {
-                          existing.hasRadiology = true;
-                        } else if (mod) {
-                          // unknown/non-empty modality — treat as non-WSI; assume radiology-capable
-                          existing.hasRadiology = true;
-                        }
-                      }
-                    }
-
-                    const studiesList = Array.from(studiesMap.values());
-                    const isExpanded = expandedCases.has(caseId);
-                    const rows: React.ReactNode[] = [];
-
-                    // Subtle zebra for case rows
-                    const isEvenStudy = idx % 2 === 0;
-                    const studyBg = isEvenStudy ? "#f5f6f7" : "#eef0f2";
-
-                    // Top-level GDC case row (no links)
-                    rows.push(
-                      <tr
-                        key={`${idx}-case`}
-                        style={{
-                          borderTop: "1px solid #eee",
-                          cursor: "pointer",
-                          background: studyBg,
-                        }}
-                        onClick={() => toggleExpanded(caseId)}
-                      >
-                        <td style={{ padding: "6px 8px" }}>{caseId}</td>
-                        <td
-                          style={{
-                            padding: "6px 8px",
-                            wordBreak: "break-all",
-                            display: "flex",
-                            alignItems: "center",
-                            gap: 8,
-                          }}
-                        >
-                          {/* show how many studies are available for this case */}
-                          <span>
-                            {studiesList.length > 0
-                              ? `${studiesList.length} study(s)`
-                              : "(/)"}
-                          </span>
-                          {/* inline expand/collapse indicator */}
-                          <span style={{ fontSize: 14 }}>
-                            {isExpanded ? "↓" : "↑"}
-                          </span>
-                        </td>
-                        {/* For non-expanded top-level rows don't show study info (leave empty) */}
-                        <td style={{ padding: "6px 8px" }}>
-                          {!isExpanded ? "" : <span>-</span>}
-                        </td>
-                        <td style={{ padding: "6px 8px" }}>
-                          {!isExpanded ? "" : <span>-</span>}
-                        </td>
-                        {/* No links on the GDC case row */}
-                        <td style={{ padding: "6px 8px" }}>
-                          <span style={{ color: "#666" }}>-</span>
-                        </td>
-                        <td style={{ padding: "6px 8px" }}>
-                          <span style={{ color: "#666" }}>-</span>
-                        </td>
-                      </tr>,
-                    );
-
-                    // When expanded, render one row per StudyInstanceUID
-                    if (isExpanded) {
-                      studiesList.forEach((study, si) => {
-                        const studyUID = study.StudyInstanceUID;
-                        const studyIdcLink =
-                          studyUID && study.hasWSI
-                            ? buildSlimStudyURL(studyUID)
-                            : null;
-                        const studyRadiologyLink =
-                          studyUID && study.hasRadiology
-                            ? CT_VIEWER_BASE +
-                              "?StudyInstanceUIDs=" +
-                              encodeURIComponent(studyUID)
-                            : null;
-
-                        // const seriesCount = study.series.length;
-                        const seriesBg = si % 2 === 0 ? "#fbfbfd" : "#f2f2f2";
-
-                        rows.push(
-                          <tr
-                            key={`${idx}-study-${si}`}
-                            style={{
-                              background: seriesBg,
-                            }}
-                          >
-                            <td style={{ padding: "6px 8px" }}>{caseId}</td>
-                            <td
-                              style={{
-                                padding: "6px 8px",
-                                wordBreak: "break-all",
-                              }}
-                            >
-                              {studyUID ?? "(/)"}
-                            </td>
-                            {/* Study-level date & description */}
-                            <td style={{ padding: "6px 8px" }}>
-                              {study.StudyDate ?? "(/)"}
-                            </td>
-                            <td
-                              style={{
-                                padding: "6px 8px",
-                                wordBreak: "break-all",
-                              }}
-                            >
-                              {study.StudyDescription ?? "(/)"}
-                            </td>
-                            <td style={{ padding: "6px 8px" }}>
-                              {studyIdcLink ? (
-                                <a
-                                  href={studyIdcLink}
-                                  target="_blank"
-                                  rel="noreferrer"
-                                  onClick={(e) => e.stopPropagation()}
-                                >
-                                  Open study
-                                </a>
-                              ) : (
-                                <span style={{ color: "#666" }}>-</span>
-                              )}
-                            </td>
-                            <td style={{ padding: "6px 8px" }}>
-                              {studyRadiologyLink ? (
-                                <a
-                                  href={studyRadiologyLink}
-                                  target="_blank"
-                                  rel="noreferrer"
-                                  onClick={(e) => e.stopPropagation()}
-                                >
-                                  Open study
-                                </a>
-                              ) : (
-                                <span style={{ color: "#666" }}>-</span>
-                              )}
-                            </td>
-                          </tr>,
-                        );
-                      });
-                    }
-
-                    return rows;
-                  })}
-                </tbody>
-              </table>
-            </>
-          )}
+          {/* Render the table view (table header + body). TableView handles empty state and uses idcImageSlice for rows */}
+          <TableView
+            mappings={mappings}
+            expandedCases={expandedCases}
+            toggleExpanded={toggleExpanded}
+            buildSlimStudyURL={buildSlimStudyURL}
+          />
         </div>
       </div>
